refactor!: Moves HisItem & HisReadRange to base package
