+++ Sources/Haystack/API/HisItem.swift
@@ -1,9 +1,7 @@
-import Haystack
-
 /// A timestamp/value pair.
 public struct HisItem {
-    let ts: DateTime
-    let val: any Val
+    public let ts: DateTime
+    public let val: any Val
     
     public init(ts: DateTime, val: any Val) {
         self.ts = ts
